Improve <app-header> resizing workaround: now it works all the time
Finally investigated what was happening: <app-header> is put into the page
with a certain width. Other elements make the desired width change and
the header is not notified.

Basically we poll every so often (lastWaitTime + 100 * log10(attemptNumber) seconds),
and tell the header to reset itself. When the width of the header matches
the width of its parent (<app-header-layout>) we do one more pass and stop.